architecture refactoring 1.0
